+++ NEWS.md
@@ -1,9 +1,15 @@
+# dbparser 1.0.2
+* Fix zip file location issue
+* Replace Secondary and third keys columns from drug framework with *other_keys* column that contains any other keys that might exist in addition to the primary key
+* Add **average-mass**, **monoisotopic-mass** and **calculated-properties** parsers.
+* Support saving parsed drugs related parsed database as csv
+
 # dbparser 1.0.1
 * Fix CRAN Note
 * Improve documentation
 * Refactor unused functions
 * Remove *Count* features from drug data set
-* Fix serveral typoes in documentation and code
+* Fix several typos in documentation and code
 * Fix consistency issue of CLASS of Data Frames Returned by dbparser
 * Check if drugbank database exist before parsing
 * Add support for *international_brands* and *salts* elements